LayoutTests/imported/w3c/web-platform-tests/css/css-writing-modes/writing-mode-parsing-sideways-lr-001-expected.txt


FAIL writing-mode: sideways-lr assert_equals: expected "sideways-lr" but got "horizontal-tb"
